+++ CHANGELOG.rst
@@ -2,6 +2,8 @@
 Changelog
 =========
 
+* Support hint, error, and fatal lines in ``git-console``.
+
 1.3.0 (2023-04-17)
 ------------------